+++ src/Region/RegionHandler.cc
@@ -335,7 +335,7 @@ bool RegionHandler::SetSpatialRequirements(
 
     std::unique_lock<std::mutex> ulock(_spatial_mutex);
     if (_region_spatial_profiles.find(region_id) == _region_spatial_profiles.end()) {
-        _region_spatial_profiles[region_id] = std::make_unique<RegionSpatialProfile>(region_id, file_id, configs);
+        _region_spatial_profiles[region_id] = std::make_shared<RegionSpatialProfile>(region_id, file_id, configs);
     } else {
         _region_spatial_profiles[region_id]->SetConfigurations(file_id, configs);
     }
@@ -484,10 +484,10 @@ void RegionHandler::UpdateNewSpectralRequirements(int region_id) {
 }
 
 bool RegionHandler::SetStatsRequirements(
-    int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetStatsRequirements_StatsConfig>& stats_configs) {
+    int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetStatsRequirements_StatsConfig>& configs) {
     // Set stats data requirements for closed region
 
-    if (stats_configs.empty() && !RegionSet(region_id)) {
+    if (configs.empty() && !RegionSet(region_id)) {
         // frontend clears requirements after region removed, prevent error in log
         return true;
     }
@@ -505,9 +505,12 @@ bool RegionHandler::SetStatsRequirements(
     // Save frame pointer
     _frames[file_id] = frame;
 
-    // Set requirements
-    ConfigId config_id(file_id, region_id);
-    _stats_req[config_id].stats_configs = stats_configs;
+    // Set configurations
+    if (_region_statistics.find(region_id) == _region_statistics.end()) {
+        _region_statistics[region_id] = std::make_unique<RegionStatistics>(region_id, file_id, configs);
+    } else {
+        _region_statistics[region_id]->SetConfigurations(file_id, configs);
+    }
     return true;
 }
 
@@ -515,51 +518,42 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
     // Clear requirements and cache for all regions or a specific region
     if (region_id == ALL_REGIONS) {
         _region_histograms.clear();
+        _region_statistics.clear();
 
         std::unique_lock<std::mutex> spatial_lock(_spatial_mutex);
         _region_spatial_profiles.clear();
         spatial_lock.unlock();
 
-        _stats_req.clear();
-
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         _spectral_req.clear();
-        spectral_lock.unlock();
-
         _spectral_cache.clear();
-        _stats_cache.clear();
+        spectral_lock.unlock();
 
         std::unique_lock pv_cut_lock(_pv_cut_mutex);
         _pv_preview_cuts.clear();
         std::unique_lock pv_cube_lock(_pv_cube_mutex);
         _pv_preview_cubes.clear();
     } else {
-        // Remove histograms for given region_id
+        // Remove region analysis for region_id
         if (_region_histograms.find(region_id) != _region_histograms.end()) {
             _region_histograms.erase(region_id);
         }
 
-        // Remove spatial profiles for given region_id
+        if (_region_statistics.find(region_id) != _region_statistics.end()) {
+            _region_statistics.erase(region_id);
+        }
+
         std::unique_lock<std::mutex> spatial_lock(_spatial_mutex);
         if (_region_spatial_profiles.find(region_id) != _region_spatial_profiles.end()) {
             _region_spatial_profiles.erase(region_id);
         }
         spatial_lock.unlock();
 
-        // Iterate through requirements and remove those for given region_id
+        // Remove spectral requirements and cache for region_id
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         for (auto it = _spectral_req.begin(); it != _spectral_req.end();) {
             if ((*it).first.region_id == region_id) {
                 it = _spectral_req.erase(it);
-            } else {
-                ++it;
-            }
-        }
-        spectral_lock.unlock();
-
-        for (auto it = _stats_req.begin(); it != _stats_req.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _stats_req.erase(it);
             } else {
                 ++it;
             }
@@ -572,16 +566,9 @@ void RegionHandler::RemoveRegionRequirementsCache(int region_id) {
                 ++it;
             }
         }
+        spectral_lock.unlock();
 
-        for (auto it = _stats_cache.begin(); it != _stats_cache.end();) {
-            if ((*it).first.region_id == region_id) {
-                it = _stats_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
-        // Do not attempt to get lock for removing temporary box region id.
+        // Remove PV preview cuts for region_id
         if (region_id > 0) {
             // Needed only for pv cut region in source image.
             std::unique_lock<std::shared_mutex> pv_cut_lock(_pv_cut_mutex);
@@ -604,33 +591,34 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
         spatial_lock.unlock();
 
         _region_histograms.clear();
+        _region_statistics.clear();
 
-        _stats_req.clear();
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         _spectral_req.clear();
-        spectral_lock.unlock();
-
         _spectral_cache.clear();
-        _stats_cache.clear();
+        spectral_lock.unlock();
 
         std::unique_lock<std::shared_mutex> pv_cut_lock(_pv_cut_mutex);
         _pv_preview_cuts.clear();
         std::unique_lock<std::shared_mutex> pv_cube_lock(_pv_cube_mutex);
         _pv_preview_cubes.clear();
     } else {
-        // Remove histograms for given file_id
+        // Remove region analysis for file_id
         for (const auto& [_, region_histogram] : _region_histograms) {
             region_histogram->ClearFileConfigsCache(file_id);
         }
 
-        // Remove spatial profiles for given file_id
+        for (const auto& [_, region_statistics] : _region_statistics) {
+            region_statistics->ClearFileConfigsCache(file_id);
+        }
+
         std::unique_lock<std::mutex> spatial_lock(_spatial_mutex);
         for (const auto& [_, spatial_profile] : _region_spatial_profiles) {
             spatial_profile->ClearFileConfigs(file_id);
         }
         spatial_lock.unlock();
 
-        // Remove requirements for given file_id
+        // Remove spectral requirements and cache for file_id
         std::unique_lock<std::mutex> spectral_lock(_spectral_mutex);
         for (auto it = _spectral_req.begin(); it != _spectral_req.end();) {
             if ((*it).first.file_id == file_id) {
@@ -639,17 +627,7 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
                 ++it;
             }
         }
-        spectral_lock.unlock();
 
-        for (auto it = _stats_req.begin(); it != _stats_req.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _stats_req.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
-        // Remove cache for given file_id
         for (auto it = _spectral_cache.begin(); it != _spectral_cache.end();) {
             if ((*it).first.file_id == file_id) {
                 it = _spectral_cache.erase(it);
@@ -657,15 +635,9 @@ void RegionHandler::RemoveFileRequirementsCache(int file_id) {
                 ++it;
             }
         }
+        spectral_lock.unlock();
 
-        for (auto it = _stats_cache.begin(); it != _stats_cache.end();) {
-            if ((*it).first.file_id == file_id) {
-                it = _stats_cache.erase(it);
-            } else {
-                ++it;
-            }
-        }
-
+        // Remove PV preview cuts and cubes for file_id
         std::unique_lock<std::shared_mutex> pv_cut_lock(_pv_cut_mutex);
         for (auto it = _pv_preview_cuts.begin(); it != _pv_preview_cuts.end();) {
             if ((*it).second->HasPreviewFileRegionIds(file_id, ALL_REGIONS)) {
@@ -698,10 +670,9 @@ void RegionHandler::ClearRegionCache(int region_id) {
             spcache.second.ClearProfiles();
         }
     }
-    for (auto& stcache : _stats_cache) {
-        if (stcache.first.region_id == region_id) {
-            stcache.second.ClearStats();
-        }
+
+    if (_region_statistics.find(region_id) != _region_statistics.end()) {
+        _region_statistics[region_id]->ClearCache();
     }
 }
 
@@ -1123,7 +1094,7 @@ bool RegionHandler::CalculatePvPreviewImage(int frame_id, int preview_id, bool q
         auto box_mask = _regions.at(box_region_id)->GetImageRegionMask(frame_id);
         RemoveRegion(box_region_id);
 
-        // Use PvPreviewCube to calculate profile with lcregion and mask
+        // Use PvPreviewCube to calculate profile with lc_region and mask
         std::vector<float> profile;
         double max_num_pixels(0.0);
         std::string message;
@@ -1864,142 +1835,66 @@ bool RegionHandler::GetRegionSpectralData(int region_id, int file_id, const Axis
     return true;
 }
 
-// ***** Fill stats data *****
-
 bool RegionHandler::FillRegionStatsData(std::function<void(CARTA::RegionStatsData stats_data)> cb, int region_id, int file_id) {
     // Fill stats data for given region and file
     if (!RegionFileIdsValid(region_id, file_id, true)) {
         return false;
     }
 
-    bool message_filled(false);
+    if ((region_id > 0) && (_region_statistics.find(region_id) == _region_statistics.end())) {
+        return false;
+    }
 
-    auto send_stats_results = [&](int region_id, int file_id, std::vector<CARTA::SetStatsRequirements_StatsConfig> stats_configs) {
-        for (auto stats_config : stats_configs) {
-            // Get stokes index
-            int stokes;
-            if (!_frames.at(file_id)->GetStokesTypeIndex(stats_config.coordinate(), stokes)) {
+    bool success(false);
+
+    for (const auto& [stats_region_id, region_statistics] : _region_statistics) {
+        if ((region_id > 0) && (stats_region_id != region_id)) {
+            continue;
+        }
+
+        auto config_file_ids = region_statistics->GetConfigFileIds(file_id);
+        if (config_file_ids.empty()) {
+            continue;
+        }
+
+        for (int stats_file_id : config_file_ids) {
+            // Get statistics for specific region id and file id
+            if (!RegionFileIdsValid(stats_region_id, stats_file_id)) {
                 continue;
             }
 
-            // Set required stats types
-            std::vector<CARTA::StatsType> required_stats;
-            for (int i = 0; i < stats_config.stats_types_size(); ++i) {
-                required_stats.push_back(stats_config.stats_types(i));
+            auto frame = _frames.at(stats_file_id);
+
+            std::vector<CARTA::SetStatsRequirements_StatsConfig> stats_configs;
+            if (!region_statistics->GetConfigurations(stats_file_id, stats_configs)) {
+                continue;
             }
 
-            // Send stats results
-            CARTA::RegionStatsData stats_message;
-            if (GetRegionStatsData(region_id, file_id, stokes, required_stats, stats_message)) {
-                cb(stats_message); // send stats data with respect to stokes
-                message_filled = true;
-            }
-        }
-    };
-
-    if (region_id > 0) {
-        // Fill stats data for specific region with file_id requirement (specific file_id or all files)
-        std::unordered_map<ConfigId, RegionStatsConfig, ConfigIdHash> region_configs = _stats_req;
-        for (auto& region_config : region_configs) {
-            if ((region_config.first.region_id == region_id) && ((region_config.first.file_id == file_id) || (file_id == ALL_FILES))) {
-                if (region_config.second.stats_configs.empty()) { // no requirements
-                    continue;
-                }
-                int config_file_id(region_config.first.file_id);
-                if (!RegionFileIdsValid(region_id, config_file_id)) { // check specific ids
-                    continue;
+            // Create data message for each configuration
+            for (auto& stats_config : stats_configs) {
+                // Get stokes and channel from config
+                int stokes(0);
+                if (!frame->GetStokesTypeIndex(stats_config.coordinate(), stokes)) {
+                    continue; // invalid image/computed Stokes
                 }
 
-                // return stats for this requirement
-                send_stats_results(region_id, config_file_id, region_config.second.stats_configs);
-            }
-        }
-    } else {
-        // (region_id < 0) Fill stats data for all regions with specific file_id requirement
-        // Find requirements with file_id
-        std::unordered_map<ConfigId, RegionStatsConfig, ConfigIdHash> region_configs = _stats_req;
-        for (auto& region_config : region_configs) {
-            if (region_config.first.file_id == file_id) {
-                if (region_config.second.stats_configs.empty()) { // no requirements
-                    continue;
-                }
-                int config_region_id(region_config.first.region_id);
-                if (!RegionFileIdsValid(config_region_id, file_id)) { // check specific ids
-                    continue;
-                }
+                // Get StokesRegion
+                int z(_frames.at(stats_file_id)->CurrentZ());
+                AxisRange z_range(z);
+                std::shared_ptr<casacore::LCRegion> lc_region;
+                StokesRegion stokes_region;
+                ApplyRegionToFile(stats_region_id, stats_file_id, z_range, stokes, lc_region, stokes_region);
+                CARTA::RegionStatsData stats_data_message;
 
-                // return stats for this requirement
-                send_stats_results(config_region_id, file_id, region_config.second.stats_configs);
+                if (region_statistics->GetRegionStatsData(stats_file_id, frame, stats_config, stokes_region, stats_data_message)) {
+                    cb(stats_data_message);
+                    success = true;
+                }
             }
         }
     }
-    return message_filled;
-}
 
-bool RegionHandler::GetRegionStatsData(
-    int region_id, int file_id, int stokes, const std::vector<CARTA::StatsType>& required_stats, CARTA::RegionStatsData& stats_message) {
-    // Fill stats message for given region, file
-    Timer t;
-
-    int z(_frames.at(file_id)->CurrentZ());
-    stokes = (stokes == CURRENT_STOKES) ? _frames.at(file_id)->CurrentStokes() : stokes;
-
-    // Start filling message
-    stats_message.set_file_id(file_id);
-    stats_message.set_region_id(region_id);
-    stats_message.set_channel(z);
-    stats_message.set_stokes(stokes);
-
-    // Check cache
-    CacheId cache_id = CacheId(file_id, region_id, stokes, z);
-    if (_stats_cache.count(cache_id)) {
-        std::map<CARTA::StatsType, double> stats_results;
-        if (_stats_cache[cache_id].GetStats(stats_results)) {
-            FillStatistics(stats_message, required_stats, stats_results);
-            return true;
-        }
-    }
-
-    // Get region
-    AxisRange z_range(z);
-    StokesRegion stokes_region;
-    std::shared_ptr<casacore::LCRegion> lc_region;
-    if (!ApplyRegionToFile(region_id, file_id, z_range, stokes, lc_region, stokes_region)) {
-        // region outside image: NaN results
-        std::map<CARTA::StatsType, double> stats_results;
-        for (const auto& carta_stat : required_stats) {
-            if (carta_stat == CARTA::StatsType::NumPixels) {
-                stats_results[carta_stat] = 0.0;
-            } else {
-                stats_results[carta_stat] = DOUBLE_NAN;
-            }
-        }
-        FillStatistics(stats_message, required_stats, stats_results);
-        // cache results
-        _stats_cache[cache_id] = StatsCache(stats_results);
-        return true;
-    }
-
-    // calculate stats
-    bool per_z(false);
-    std::map<CARTA::StatsType, std::vector<double>> stats_map;
-    if (_frames.at(file_id)->GetRegionStats(stokes_region, required_stats, per_z, stats_map)) {
-        // convert vector to single value in map
-        std::map<CARTA::StatsType, double> stats_results;
-        for (auto& value : stats_map) {
-            stats_results[value.first] = value.second[0];
-        }
-
-        // add values to message
-        FillStatistics(stats_message, required_stats, stats_results);
-        // cache results
-        _stats_cache[cache_id] = StatsCache(stats_results);
-
-        spdlog::performance("Fill region stats in {:.3f} ms", t.Elapsed().ms());
-        return true;
-    }
-
-    return false;
+    return success;
 }
 
 bool RegionHandler::FillSpatialProfileData(std::function<void(CARTA::SpatialProfileData profile_data)> cb, int file_id, int region_id) {