+++ src/Region/RegionHandler.h
@@ -20,6 +20,7 @@
 #include "Region.h"
 #include "RegionAnalysis/RegionHistogram.h"
 #include "RegionAnalysis/RegionSpatialProfile.h"
+#include "RegionAnalysis/RegionStatistics.h"
 
 namespace carta {
 
@@ -49,8 +50,8 @@ public:
         int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetSpatialRequirements_SpatialConfig>& configs);
     bool SetSpectralRequirements(int region_id, int file_id, std::shared_ptr<Frame> frame,
         const std::vector<CARTA::SetSpectralRequirements_SpectralConfig>& configs);
-    bool SetStatsRequirements(int region_id, int file_id, std::shared_ptr<Frame> frame,
-        const std::vector<CARTA::SetStatsRequirements_StatsConfig>& stats_configs);
+    bool SetStatsRequirements(
+        int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetStatsRequirements_StatsConfig>& configs);
 
     // Calculations
     bool FillRegionHistogramData(std::function<void(CARTA::RegionHistogramData histogram_data)> cb, int region_id, int file_id);
@@ -69,7 +70,7 @@ public:
     bool IsClosedRegion(int region_id);
 
     // Generate PV image or preview image
-    bool CalculatePvImage(const CARTA::PvRequest& pv_request, std::shared_ptr<Frame>& frame, GeneratorProgressCallback progress_callback,
+    bool CalculatePvImage(const CARTA::PvRequest& pv_request, std::shared_ptr<Frame> frame, GeneratorProgressCallback progress_callback,
         CARTA::PvResponse& pv_response, GeneratedImage& pv_image);
     // Update PV preview image when PV cut region (region_id) changes
     bool UpdatePvPreviewRegion(int region_id, RegionState& region_state);
@@ -103,17 +104,12 @@ private:
     void ClearRegionCache(int region_id);
 
     // Apply region to image
-    std::shared_ptr<casacore::LCRegion> ApplyRegionToFile(
-        int region_id, int file_id, const StokesSource& stokes_source = StokesSource(), bool report_error = true);
-    bool ApplyRegionToFile(int region_id, int file_id, const AxisRange& z_range, int stokes, std::shared_ptr<casacore::LCRegion> lc_region,
-        StokesRegion& stokes_region);
+
 
     // Data stream helpers
     bool GetRegionSpectralData(int region_id, int file_id, const AxisRange& z_range, std::string& coordinate, int stokes_index,
         std::vector<CARTA::StatsType>& required_stats, bool report_error,
         const std::function<void(std::map<CARTA::StatsType, std::vector<double>>, float)>& partial_results_callback);
-    bool GetRegionStatsData(
-        int region_id, int file_id, int stokes, const std::vector<CARTA::StatsType>& required_stats, CARTA::RegionStatsData& stats_message);
     bool GetLineSpatialData(int file_id, int region_id, const std::string& coordinate, int stokes_index, int width,
         const std::function<void(std::vector<float>&, casacore::Quantity&)>& spatial_profile_callback);
 
@@ -164,21 +160,14 @@ private:
 
     // Region analysis
     std::unordered_map<int, std::unique_ptr<RegionHistogram>> _region_histograms;
-    std::unordered_map<int, std::unique_ptr<RegionSpatialProfile>> _region_spatial_profiles;
+    std::unordered_map<int, std::shared_ptr<RegionSpatialProfile>> _region_spatial_profiles;
+    std::unordered_map<int, std::unique_ptr<RegionStatistics>> _region_statistics;
     std::mutex _spatial_mutex;
 
-    // Requirements; ConfigId key contains file, region
+    // Spectral profile requirements, cache, and list of supported stats
     std::unordered_map<ConfigId, RegionSpectralConfig, ConfigIdHash> _spectral_req;
-    std::unordered_map<ConfigId, RegionStatsConfig, ConfigIdHash> _stats_req;
-
-    // Lock to add/remove requirements
-    std::mutex _spectral_mutex;
-
-    // Cache; CacheId key contains file, region, stokes, (optional) z index
     std::unordered_map<CacheId, SpectralCache, CacheIdHash> _spectral_cache;
-    std::unordered_map<CacheId, StatsCache, CacheIdHash> _stats_cache;
-
-    // Spectral profiles to calculate with ImageStatistics.
+    std::mutex _spectral_mutex;
     std::vector<CARTA::StatsType> _spectral_stats = {CARTA::StatsType::Sum, CARTA::StatsType::FluxDensity, CARTA::StatsType::Mean,
         CARTA::StatsType::RMS, CARTA::StatsType::Sigma, CARTA::StatsType::SumSq, CARTA::StatsType::Min, CARTA::StatsType::Max,
         CARTA::StatsType::Extrema, CARTA::StatsType::NumPixels};